added value to users

+++ db/migrate/20170220104054_create_bg_tests.rb
@@ -1,7 +1,7 @@
 class CreateBgTests < ActiveRecord::Migration[5.0]
   def change
     create_table :bg_tests do |t|
-      t.decimal(3,1) :bg_result
+      t.float :bg_result
       t.date :date
       t.time :time
       t.references :user, foreign_key: true
